spring-boot-tests/pyproject.toml
[project]
name = "spring-boot-tests"
version = "1.0.0"
description = "Python тесты для Spring Boot приложения"
requires-python = ">=3.11"

dependencies = [
    "requests>=2.31.0",
    "httpx>=0.24.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
test = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-xdist>=3.3.0",
    "pytest-wiremock>=1.0.0",
    "wiremock>=2.0.0",
    "allure-pytest>=2.13.0",
    "pytest-html>=4.0.0",
    "pytest-xdist>=3.3.0",
    "faker>=19.0.0",
    "deepdiff>=6.3.0",
    "jsonschema>=4.19.0",
    "jsonpath-ng>=1.5.3",
    "testcontainers>=3.8.0",
    "testcontainers-postgres>=3.8.0",
    "testcontainers-redis>=3.8.0",
    "prance>=0.22.0",
    "openapi-spec-validator>=0.5.0",
]

dev = [
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
    "pre-commit>=3.5.0",
]

performance = [
    "locust>=2.15.0",
]

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--tb=short",
    "--junit-xml=reports/junit.xml",
    "--html=reports/report.html",
    "--self-contained-html",
]
markers = [
    "smoke: Smoke tests",
    "api: API tests",
    "integration: Integration tests",
    "contract: Contract tests",
    "performance: Performance tests",
    "positive: Positive test scenarios",
    "negative: Negative test scenarios",
]

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
